Clang: ignore warning that checks doxygen comments (-Wdocumentation)
If -Werror and -Wdocumentation are both enabled, then clang is very unhappy with some of the Doxygen tags in the comments (eg, "@sa @ref" and some @tparam that are embedded in comments that do not actually attach to definitions with such template arguments). For now we should just disable this warning.